requirements.txt
Flask
google-generativeai
SpeechRecognition
gTTS
langdetect
